Implement hashCode method in IdentityInfo
HashCode is now calculated based on file paths of the keyfiles
instead of file objects because two file objects are different
even if they point to the same keyfile.
This was done to avoid creating redundant sftp connections because
it was not possible to identify that there already exist a sftp
connection with same properties.
Fixes wso2/product-ei#3384